Create Dataset_v1.2-DRAFT.json
Add about as recommended.

In Bioschemas, following what is done in RO-Crate, about should be used to specified the nature of the content of the dataset, uisng DefinedTerm whenever possible. For instance, if a dataset contains information about mice, the about could be ncit:10090. Ideally this info should go as part of the Bioschemas description but currently there is no way to capture those. 

This change is discussed in https://github.com/BioSchemas/specifications/issues/629